pyproject.toml
[project]
name = "day4-blog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.16.5",
    "bcrypt==4.0.1",
    "email-validator>=2.2.0",
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.12.5",
    "pymysql>=1.1.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.44.0",
]
